test(e2e): split large test files for improved parallelization
Split two large E2E test files into smaller, focused modules to enable
better parallel execution with Playwright workers.

Changes:
- Split category-management.spec.ts (1,401 lines) into 4 files:
  - category-management-crud.spec.ts (CRUD operations)
  - category-management-integration.spec.ts (prompt system integration)
  - category-management-ui.spec.ts (navigation and display)
  - category-management-accessibility.spec.ts (keyboard and a11y)

- Split usage-counter-sorting.spec.ts (914 lines) into 3 files:
  - usage-counter-sorting-ui.spec.ts (UI display and behavior)
  - usage-counter-sorting-logic.spec.ts (cross-context sync and migration)
  - usage-counter-sorting-edge-cases.spec.ts (edge cases and performance)

Benefits:
- Eliminates bottlenecks with 4-worker parallelization
- Expected 60-75% faster E2E execution (4-5min → 1.5-2min)
- Improved maintainability with smaller, focused test files
- All files now 300-500 lines (optimal for parallel execution)

Test Results: ✅ 69 tests passed in 1m 4.5s with 4 workers
